Create baklava syrup recipe page and associated links
Created baklava-syrup.html with links to and from index.html.
Page contains image, syrup.jpg, which is located in the img directory.
Noticed inconsistent style in tzatziki.html: Description does not use
ampersand but instead uses the word 'and'. This is inconsistent with the
styling of the rest of the website and will need to be rectified in a
dedicated, future commit.