Styling of tzatziki.html and button creation
Styled tzatziki.html. style.css is now set up for all recipe pages,
all they will need is the right classes declared and link created for the stylesheet
to apply.
Created a 'button' class for recipe links and for the home link. Styled the buttons
in index.html.

Changes to be committed:
	modified:   index.html
	modified:   recipes/tzatziki.html
	modified:   style.css